requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
cycler==0.12.1
fonttools==4.43.1
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
threadpoolctl==3.2.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
